copy_file: Apply fixes after review
Limit the number of descriptors placed in a queue to NUM_BUFF (8)
to stop it from growing indefinitelly.
Add condition variables to avoid busy looping when queue is empty or
full. Change the read mutex to unique_lock to open mutex when waiting
for the write thread to pop a buffer when queue is full.
Use std::vector in Descriptor class to avoid mamcopy.
Add getters to Descriptor class ang change fields to private.

Signed-off-by: mplantykow <m.plantykow@gmail.com>